+++ frontend/src/components/GarageAdmin/EditSection.vue
