+++ frontend/src/components/GarageAdmin/HoursSection.vue
