+++ frontend/src/components/GarageAdmin/InfoSection.vue
