+++ frontend/src/components/GarageAdmin/PhotosSection.vue
