+++ frontend/src/components/GarageAdmin/SearchBar.vue
