+++ frontend/src/views/GarageAdmin/EditView.vue
@@ -1,10 +1,10 @@
 <script setup lang="ts">
-import PageHead from './_components/PageHead.vue';
-import EditSection from './_components/EditSection.vue';
-import InfoSection from './_components/InfoSection.vue';
-import HoursSection from './_components/HoursSection.vue';
-import PhotosSection from './_components/PhotosSection.vue';
-import SubscriptionSection from './_components/SubscriptionSection.vue';
+import PageHead from '@/components/GarageAdmin/PageHead.vue';
+import EditSection from '@/components/GarageAdmin/EditSection.vue';
+import InfoSection from '@/components/GarageAdmin/InfoSection.vue';
+import HoursSection from '@/components/GarageAdmin/HoursSection.vue';
+import PhotosSection from '@/components/GarageAdmin/PhotosSection.vue';
+import SubscriptionSection from '@/components/GarageAdmin/SubscriptionSection.vue';
 </script>
 
 <template>
