+++ frontend/src/views/GarageAdmin/OverView.vue
@@ -1,12 +1,16 @@
 <script setup lang="ts">
 import { ref } from 'vue';
-import OrderCard from './_components/OrderCard.vue';
-import OrderDetailModal from './_components/OrderDetailModal.vue';
-import PageHead from './_components/PageHead.vue';
-import SearchBar from './_components/SearchBar.vue';
-import type { Order } from './types';
+import OrderCard from '@/components/GarageAdmin/OrderCard.vue';
+import OrderDetailModal from '@/components/GarageAdmin/OrderDetailModal.vue';
+import type { Order } from '@/components/GarageAdmin/types';
 
 // 假資料
+const stats = [
+  { label: '今日預約', value: 12, icon: 'event_note' },
+  { label: '待確認工單', value: 5, icon: 'assignment' },
+  { label: '本月完成', value: 45, icon: 'trending_up' },
+];
+
 const orders = ref<Order[]>([
   {
     id: 'ORD-20260105-001',
@@ -72,7 +76,6 @@ const orders = ref<Order[]>([
 
 const showModal = ref(false);
 const selectedOrder = ref<Order | null>(null);
-const keyword = ref('');
 
 const open = (order: Order) => {
   selectedOrder.value = order;
@@ -86,12 +89,35 @@ const save = () => {
 
 <template>
   <div class="space-y-6">
-    <PageHead title="工單管理" subtitle="系統工單配對、查詢" />
-    <SearchBar placeholder="搜尋工單號、客戶姓名、電話或車牌..." v-model:keyword="keyword" />
+    <!-- 標題 -->
+    <div>
+      <h1 class="text-2xl font-semibold text-[#4a4a43]">總覽</h1>
+      <p class="mt-1 text-[#8a8a7d]">今日維修廠營運概況</p>
+    </div>
 
-    <!-- 工單列表 -->
-    <div class="space-y-6">
-      <OrderCard v-for="order in orders" :key="order.id" :order="order" @open="open" />
+    <!-- 統計卡片 -->
+    <div class="grid grid-cols-1 md:grid-cols-3 gap-6">
+      <div
+        v-for="item in stats"
+        :key="item.label"
+        class="bg-[#f5f4f0] rounded-2xl p-6 flex items-center justify-between shadow-sm"
+      >
+        <div>
+          <p class="text-[#8a8a7d] text-sm mb-1">{{ item.label }}</p>
+          <p class="text-3xl font-semibold text-[#4a4a43]">{{ item.value }}</p>
+        </div>
+        <div class="w-12 h-12 flex items-center justify-center text-xl bg-white rounded-xl">
+          <span class="material-symbols-outlined">{{ item.icon }}</span>
+        </div>
+      </div>
+    </div>
+
+    <div>
+      <h2 class="text-xl font-semibold text-[#4a4a43]">近期工單</h2>
+      <!-- 工單列表 -->
+      <section class="mt-5 space-y-6">
+        <OrderCard v-for="order in orders" :key="order.id" :order="order" @open="open" />
+      </section>
     </div>
   </div>
